Multiverse + Hyperverse are now soft dependencies. should resume fill progress upon restart.

+++ src/main/java/com/wimbli/WorldBorder/cmd/CmdRadius.java
@@ -34,8 +34,8 @@ public class CmdRadius extends WBCmd {
 
         double x = border.getX();
         double z = border.getZ();
-        int radiusX;
-        int radiusZ;
+        double radiusX;
+        double radiusZ;
         try {
             if (params.get(0).startsWith("+")) {
                 // Add to the current radius
